requirements.txt
hidapi>=0.14.0
plyer>=2.1.0
pystray>=0.19.5
Pillow>=10.0.0
winotify>=1.1.0
